feat: add custom materials and map to values (#220)
* feat: add function to insert mask at front of material stack

* feat: directional process rate based

* feat: material value mapping in single particle process rates

* feat: single particle process python bindings updated

* feat: add mapping in IsotropicProcess

* refactor: material mapping, allow custom material registration

* feat: materials on CUDA device code

* fix: plasma etching models

* feat: add material factors in plasma etching model

* test adding custom material to domain

* refactor: use singleton MaterialRegistry

* material registry tests

* refactor: update directional process

* refactor: udpate ion beam etching model

* fix: faraday cage model, update python bindings

* small map refactor

* fix: bug in mask blur GDS

* feat: parallel blur loop

* fix: correct material check function in CSV file process

* refactor: material info

* feat: fix build

* chore: format

* fix: material info requires string for custom values

+++ tests/fluxEngineBenchmarks/plasmaEtching.cpp
@@ -28,10 +28,10 @@ int main(int argc, char *argv[]) {
 
     // use pre-defined model SF6O2 etching model
     auto modelParams = SF6O2Etching<NumericType, D>::defaultParameters();
-    modelParams.beta_E[static_cast<int>(Material::Si)] = 1.0;
-    modelParams.beta_E[static_cast<int>(Material::Mask)] = 1.0;
-    modelParams.beta_P[static_cast<int>(Material::Si)] = 1.0;
-    modelParams.beta_P[static_cast<int>(Material::Mask)] = 1.0;
+    modelParams.beta_E[Material::Si] = 1.0;
+    modelParams.beta_E[Material::Mask] = 1.0;
+    modelParams.beta_P[Material::Si] = 1.0;
+    modelParams.beta_P[Material::Mask] = 1.0;
 
     modelParams.ionFlux = 1.;
     modelParams.etchantFlux = 4.5e3;
